requirements.txt
Flask==3.0.3
google-generativeai==0.7.0
pypdf2==3.0.1
pandas==2.1.4
openpyxl==3.1.2
python-dotenv==1.0.1
